Added code to prevent routing for users without logging.
