+++ src/components/globalHeader/GlobalHeader.jsx
@@ -1,4 +1,4 @@
-import { useContext } from "react";
+import {useContext, useEffect} from "react";
 import {Link, useNavigate} from "react-router-dom";
 import { GlobalContext } from "../../context/globalContext";
 import styles from "./GlobalHeader.module.css";
@@ -6,6 +6,18 @@ import styles from "./GlobalHeader.module.css";
 export function GlobalHeader(props) {
     const context = useContext(GlobalContext);
     const navigate = useNavigate();
+
+    const getRoutByUserType = () => {
+        // return context.userType === "admin" ? "/users" : "/contacts";
+        return "/contacts";
+    };
+
+    useEffect(() => {
+        context.isLoggedIn
+            ? navigate(getRoutByUserType(), { replace: true })
+            : navigate("/login", { replace: true });
+    }, []);
+
     const logoutUser = () => {
         context.logoutUser();
         navigate("/login", { replace: true });
